+++ src/adt_ai/cli.py
@@ -4,6 +4,7 @@ from adt_ai.cli_startup import print_startup_failure as _print_startup_failure
 
 try:
     from adt_ai import cli_commands_exports as _commands_exports
+    from adt_ai import cli_commands_flow as _commands_flow
     from adt_ai import cli_commands_history as _commands_history
     from adt_ai import cli_commands_recompile as _commands_recompile
     from adt_ai import cli_constants as _constants
@@ -28,6 +29,7 @@ else:
         _context,
         _commands_history,
         _commands_recompile,
+        _commands_flow,
         _commands_exports,
         _runtime,
     )